Move subquery handling into its own file.
